requirements.txt
python-dotenv==1.0.1

# ==== Core LangChain và Dependencies (Đã sửa lỗi version) ====
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# Các thư viện tích hợp (Bắt buộc phải cài thêm ở phiên bản mới)
langchain-google-genai>=1.0.3
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1

# Google GenAI (Cập nhật để khớp với langchain-google-genai)
google-generativeai>=0.5.0
protobuf>=4.25.3

# ==== AI/ML (Giữ nguyên lựa chọn của bạn) ====
torch>=2.2.2
transformers>=4.37.2
sentence-transformers>=2.2.2

# ==== Vector store & embeddings ====
faiss-cpu>=1.7.4
chromadb>=0.5.0

# ==== Web & API ====
Flask==3.0.0
uvicorn==0.27.1
fastapi==0.110.0
requests==2.31.0

# ==== STT/TTS ====
faster-whisper>=0.10.0
edge-tts>=6.1.9

# ==== Utility ====
tqdm==4.66.2
sqlalchemy==2.0.25
coloredlogs==15.0.1